hassio-google-drive-backup/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
requests
python-dateutil==2.6.1
CherryPy
colored
watchdog
pyyaml
dnspython
